+++ execution-engine/common/src/key.rs
@@ -1,61 +1,30 @@
 use core::fmt::Write;
 
-use super::alloc::vec::Vec;
-use super::bytesrepr::{Error, FromBytes, ToBytes, N32, OPTION_SIZE, U32_SIZE};
+use crate::alloc::vec::Vec;
+use crate::bytesrepr::{Error, FromBytes, ToBytes, N32, U32_SIZE};
 use crate::contract_api::pointers::*;
-use bitflags;
+use crate::uref::{URef, UREF_SIZE_SERIALIZED};
 
-bitflags! {
-    #[allow(clippy::derive_hash_xor_eq)]
-    pub struct AccessRights: u8 {
-        const READ  = 0b001;
-        const WRITE = 0b010;
-        const ADD   = 0b100;
-        const READ_ADD       = Self::READ.bits | Self::ADD.bits;
-        const READ_WRITE     = Self::READ.bits | Self::WRITE.bits;
-        const ADD_WRITE      = Self::ADD.bits  | Self::WRITE.bits;
-        const READ_ADD_WRITE = Self::READ.bits | Self::ADD.bits | Self::WRITE.bits;
-    }
-}
-
-impl AccessRights {
-    pub fn is_readable(self) -> bool {
-        self & AccessRights::READ == AccessRights::READ
-    }
-
-    pub fn is_writeable(self) -> bool {
-        self & AccessRights::WRITE == AccessRights::WRITE
-    }
-
-    pub fn is_addable(self) -> bool {
-        self & AccessRights::ADD == AccessRights::ADD
-    }
-}
-
-impl core::fmt::Display for AccessRights {
-    fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
-        match *self {
-            AccessRights::READ => write!(f, "READ"),
-            AccessRights::WRITE => write!(f, "WRITE"),
-            AccessRights::ADD => write!(f, "ADD"),
-            AccessRights::READ_ADD => write!(f, "READ_ADD"),
-            AccessRights::READ_WRITE => write!(f, "READ_WRITE"),
-            AccessRights::ADD_WRITE => write!(f, "ADD_WRITE"),
-            AccessRights::READ_ADD_WRITE => write!(f, "READ_ADD_WRITE"),
-            _ => write!(f, "UNKNOWN"),
-        }
-    }
-}
+const ACCOUNT_ID: u8 = 0;
+const HASH_ID: u8 = 1;
+const UREF_ID: u8 = 2;
+const LOCAL_ID: u8 = 3;
 
 pub const LOCAL_SEED_SIZE: usize = 32;
 pub const LOCAL_KEY_HASH_SIZE: usize = 32;
+
+const KEY_ID_SIZE: usize = 1; // u8 used to determine the ID
+const ACCOUNT_KEY_SIZE: usize = KEY_ID_SIZE + U32_SIZE + N32;
+const HASH_KEY_SIZE: usize = KEY_ID_SIZE + U32_SIZE + N32;
+pub const UREF_SIZE: usize = KEY_ID_SIZE + UREF_SIZE_SERIALIZED;
+const LOCAL_SIZE: usize = KEY_ID_SIZE + U32_SIZE + LOCAL_SEED_SIZE + U32_SIZE + LOCAL_KEY_HASH_SIZE;
 
 #[repr(C)]
 #[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Hash)]
 pub enum Key {
     Account([u8; 32]),
     Hash([u8; 32]),
-    URef([u8; 32], Option<AccessRights>), //TODO: more bytes?
+    URef(URef),
     Local {
         seed: [u8; LOCAL_SEED_SIZE],
         key_hash: [u8; LOCAL_KEY_HASH_SIZE],
@@ -77,10 +46,15 @@ impl core::fmt::Display for Key {
         match self {
             Key::Account(addr) => write!(f, "Account({})", addr_to_hex(addr)),
             Key::Hash(addr) => write!(f, "Hash({})", addr_to_hex(addr)),
-            Key::URef(addr, Some(access_rights)) => {
-                write!(f, "URef({}, {})", addr_to_hex(addr), access_rights)
+            Key::URef(uref) => {
+                let addr = uref.addr();
+                let access_rights_o = uref.access_rights();
+                if let Some(access_rights) = access_rights_o {
+                    write!(f, "URef({}, {})", addr_to_hex(&addr), access_rights)
+                } else {
+                    write!(f, "URef({}, None)", addr_to_hex(&addr))
+                }
             }
-            Key::URef(addr, None) => write!(f, "URef({}, None)", addr_to_hex(addr)),
             Key::Local { seed, key_hash } => {
                 write!(f, "Local({}, {})", addr_to_hex(seed), addr_to_hex(key_hash))
             }
@@ -95,12 +69,11 @@ impl core::fmt::Debug for Key {
 }
 
 use alloc::string::String;
-use Key::*;
 
 impl Key {
     pub fn to_u_ptr<T>(self) -> Option<UPointer<T>> {
-        if let URef(id, Some(rights)) = self {
-            Some(UPointer::new(id, rights))
+        if let Key::URef(uref) = self {
+            UPointer::from_uref(uref).ok()
         } else {
             None
         }
@@ -108,8 +81,8 @@ impl Key {
 
     pub fn to_c_ptr(self) -> Option<ContractPointer> {
         match self {
-            URef(id, Some(rights)) => Some(ContractPointer::URef(UPointer::new(id, rights))),
-            Hash(id) => Some(ContractPointer::Hash(id)),
+            Key::URef(uref) => UPointer::from_uref(uref).map(ContractPointer::URef).ok(),
+            Key::Hash(id) => Some(ContractPointer::Hash(id)),
             _ => None,
         }
     }
@@ -117,71 +90,41 @@ impl Key {
     /// Returns bytes of an account
     pub fn as_account(&self) -> Option<[u8; 32]> {
         match self {
-            Account(bytes) => Some(*bytes),
+            Key::Account(bytes) => Some(*bytes),
             _ => None,
         }
     }
 
     pub fn normalize(self) -> Key {
         match self {
-            Key::URef(id, _) => Key::URef(id, None),
+            Key::URef(uref) => Key::URef(uref.remove_access_rights()),
             other => other,
         }
-    }
-}
-
-const ACCOUNT_ID: u8 = 0;
-const HASH_ID: u8 = 1;
-const UREF_ID: u8 = 2;
-const LOCAL_ID: u8 = 3;
-
-const KEY_ID_SIZE: usize = 1; // u8 used to determine the ID
-const ACCESS_RIGHTS_SIZE: usize = 1; // u8 used to tag AccessRights
-const ACCOUNT_KEY_SIZE: usize = KEY_ID_SIZE + U32_SIZE + N32;
-const HASH_KEY_SIZE: usize = KEY_ID_SIZE + U32_SIZE + N32;
-pub const UREF_SIZE: usize = KEY_ID_SIZE + U32_SIZE + N32 + OPTION_SIZE + ACCESS_RIGHTS_SIZE;
-const LOCAL_SIZE: usize = KEY_ID_SIZE + U32_SIZE + LOCAL_SEED_SIZE + U32_SIZE + LOCAL_KEY_HASH_SIZE;
-
-impl ToBytes for AccessRights {
-    fn to_bytes(&self) -> Result<Vec<u8>, Error> {
-        self.bits.to_bytes()
-    }
-}
-
-impl FromBytes for AccessRights {
-    fn from_bytes(bytes: &[u8]) -> Result<(Self, &[u8]), Error> {
-        let (id, rest): (u8, &[u8]) = FromBytes::from_bytes(bytes)?;
-        let access_rights = match AccessRights::from_bits(id) {
-            Some(rights) => Ok(rights),
-            None => Err(Error::FormattingError),
-        };
-        access_rights.map(|rights| (rights, rest))
     }
 }
 
 impl ToBytes for Key {
     fn to_bytes(&self) -> Result<Vec<u8>, Error> {
         match self {
-            Account(addr) => {
+            Key::Account(addr) => {
                 let mut result = Vec::with_capacity(ACCOUNT_KEY_SIZE);
                 result.push(ACCOUNT_ID);
                 result.append(&mut addr.to_bytes()?);
                 Ok(result)
             }
-            Hash(hash) => {
+            Key::Hash(hash) => {
                 let mut result = Vec::with_capacity(HASH_KEY_SIZE);
                 result.push(HASH_ID);
                 result.append(&mut hash.to_bytes()?);
                 Ok(result)
             }
-            URef(rf, maybe_access_rights) => {
+            Key::URef(uref) => {
                 let mut result = Vec::with_capacity(UREF_SIZE);
                 result.push(UREF_ID);
-                result.append(&mut rf.to_bytes()?);
-                result.append(&mut maybe_access_rights.to_bytes()?);
+                result.append(&mut uref.to_bytes()?);
                 Ok(result)
             }
-            Local { seed, key_hash } => {
+            Key::Local { seed, key_hash } => {
                 let mut result = Vec::with_capacity(LOCAL_SIZE);
                 result.push(LOCAL_ID);
                 result.append(&mut seed.to_bytes()?);
@@ -198,22 +141,20 @@ impl FromBytes for Key {
         match id {
             ACCOUNT_ID => {
                 let (addr, rem): ([u8; 32], &[u8]) = FromBytes::from_bytes(rest)?;
-                Ok((Account(addr), rem))
+                Ok((Key::Account(addr), rem))
             }
             HASH_ID => {
                 let (hash, rem): ([u8; 32], &[u8]) = FromBytes::from_bytes(rest)?;
-                Ok((Hash(hash), rem))
+                Ok((Key::Hash(hash), rem))
             }
             UREF_ID => {
-                let (rf, rem): ([u8; 32], &[u8]) = FromBytes::from_bytes(rest)?;
-                let (maybe_access_rights, rem2): (Option<AccessRights>, &[u8]) =
-                    FromBytes::from_bytes(rem)?;
-                Ok((URef(rf, maybe_access_rights), rem2))
+                let (uref, rem): (URef, &[u8]) = FromBytes::from_bytes(rest)?;
+                Ok((Key::URef(uref), rem))
             }
             LOCAL_ID => {
                 let (seed, rest): ([u8; 32], &[u8]) = FromBytes::from_bytes(rest)?;
                 let (key_hash, rest): ([u8; 32], &[u8]) = FromBytes::from_bytes(rest)?;
-                Ok((Local { seed, key_hash }, rest))
+                Ok((Key::Local { seed, key_hash }, rest))
             }
             _ => Err(Error::FormattingError),
         }
@@ -253,8 +194,8 @@ impl ToBytes for Vec<Key> {
 #[allow(clippy::unnecessary_operation)]
 #[cfg(test)]
 mod tests {
-    use crate::key::AccessRights;
     use crate::key::Key;
+    use crate::uref::{AccessRights, URef};
     use alloc::string::String;
 
     fn test_readable(right: AccessRights, is_true: bool) {
@@ -311,7 +252,7 @@ mod tests {
             format!("{}", account_key),
             format!("Account({})", expected_hash)
         );
-        let uref_key = Key::URef(addr_array, Some(AccessRights::READ));
+        let uref_key = Key::URef(URef::new(addr_array, AccessRights::READ));
         assert_eq!(
             format!("{}", uref_key),
             format!("URef({}, READ)", expected_hash)