+++ execution-engine/engine/src/execution.rs
@@ -1,5 +1,6 @@
 use std::cell::RefCell;
 use std::collections::{BTreeMap, HashMap, HashSet};
+use std::convert::TryFrom;
 use std::fmt;
 use std::iter::IntoIterator;
 use std::rc::Rc;
@@ -16,7 +17,8 @@ use wasmi::{
 };
 
 use common::bytesrepr::{deserialize, Error as BytesReprError, ToBytes};
-use common::key::{AccessRights, Key};
+use common::key::Key;
+use common::uref::AccessRights;
 use common::value::Value;
 use shared::newtypes::{CorrelationId, Validated};
 use shared::transform::TypeMismatch;
@@ -24,13 +26,7 @@ use storage::global_state::StateReader;
 
 use args::Args;
 use engine_state::execution_result::ExecutionResult;
-use functions::{
-    ADD_FUNC_INDEX, ADD_UREF_FUNC_INDEX, CALL_CONTRACT_FUNC_INDEX, GAS_FUNC_INDEX,
-    GET_ARG_FUNC_INDEX, GET_CALL_RESULT_FUNC_INDEX, GET_FN_FUNC_INDEX, GET_READ_FUNC_INDEX,
-    GET_UREF_FUNC_INDEX, HAS_UREF_FUNC_INDEX, IS_VALID_FN_INDEX, LOAD_ARG_FUNC_INDEX,
-    NEW_FUNC_INDEX, PROTOCOL_VERSION_FUNC_INDEX, READ_FUNC_INDEX, RET_FUNC_INDEX,
-    REVERT_FUNC_INDEX, SEED_FN_INDEX, SER_FN_FUNC_INDEX, STORE_FN_INDEX, WRITE_FUNC_INDEX,
-};
+use function_index::FunctionIndex;
 use resolvers::create_module_resolver;
 use resolvers::error::ResolverError;
 use resolvers::memory_resolver::MemoryResolver;
@@ -455,8 +451,9 @@ where
         index: usize,
         args: RuntimeArgs,
     ) -> Result<Option<RuntimeValue>, Trap> {
-        match index {
-            READ_FUNC_INDEX => {
+        let func = FunctionIndex::try_from(index).expect("unknown function index");
+        match func {
+            FunctionIndex::ReadFuncIndex => {
                 // args(0) = pointer to key in Wasm memory
                 // args(1) = size of key in Wasm memory
                 let (key_ptr, key_size) = Args::parse(args)?;
@@ -464,7 +461,7 @@ where
                 Ok(Some(RuntimeValue::I32(size as i32)))
             }
 
-            SER_FN_FUNC_INDEX => {
+            FunctionIndex::SerFnFuncIndex => {
                 // args(0) = pointer to name in Wasm memory
                 // args(1) = size of name in Wasm memory
                 let (name_ptr, name_size) = Args::parse(args)?;
@@ -472,7 +469,7 @@ where
                 Ok(Some(RuntimeValue::I32(size as i32)))
             }
 
-            WRITE_FUNC_INDEX => {
+            FunctionIndex::WriteFuncIndex => {
                 // args(0) = pointer to key in Wasm memory
                 // args(1) = size of key
                 // args(2) = pointer to value
@@ -482,7 +479,7 @@ where
                 Ok(None)
             }
 
-            ADD_FUNC_INDEX => {
+            FunctionIndex::AddFuncIndex => {
                 // args(0) = pointer to key in Wasm memory
                 // args(1) = size of key
                 // args(2) = pointer to value
@@ -492,7 +489,7 @@ where
                 Ok(None)
             }
 
-            NEW_FUNC_INDEX => {
+            FunctionIndex::NewFuncIndex => {
                 // args(0) = pointer to key destination in Wasm memory
                 // args(1) = pointer to initial value
                 // args(2) = size of initial value
@@ -501,35 +498,35 @@ where
                 Ok(None)
             }
 
-            GET_READ_FUNC_INDEX => {
+            FunctionIndex::GetReadFuncIndex => {
                 // args(0) = pointer to destination in Wasm memory
                 let dest_ptr = Args::parse(args)?;
                 self.set_mem_from_buf(dest_ptr)?;
                 Ok(None)
             }
 
-            GET_FN_FUNC_INDEX => {
+            FunctionIndex::GetFnFuncIndex => {
                 // args(0) = pointer to destination in Wasm memory
                 let dest_ptr = Args::parse(args)?;
                 self.set_mem_from_buf(dest_ptr)?;
                 Ok(None)
             }
 
-            LOAD_ARG_FUNC_INDEX => {
+            FunctionIndex::LoadArgFuncIndex => {
                 // args(0) = index of host runtime arg to load
                 let i = Args::parse(args)?;
                 let size = self.load_arg(i)?;
                 Ok(Some(RuntimeValue::I32(size as i32)))
             }
 
-            GET_ARG_FUNC_INDEX => {
+            FunctionIndex::GetArgFuncIndex => {
                 // args(0) = pointer to destination in Wasm memory
                 let dest_ptr = Args::parse(args)?;
                 self.set_mem_from_buf(dest_ptr)?;
                 Ok(None)
             }
 
-            RET_FUNC_INDEX => {
+            FunctionIndex::RetFuncIndex => {
                 // args(0) = pointer to value
                 // args(1) = size of value
                 // args(2) = pointer to extra returned urefs
@@ -544,7 +541,7 @@ where
                 ))
             }
 
-            CALL_CONTRACT_FUNC_INDEX => {
+            FunctionIndex::CallContractFuncIndex => {
                 // args(0) = pointer to key where contract is at in global state
                 // args(1) = size of key
                 // args(2) = pointer to function arguments in Wasm memory
@@ -567,14 +564,14 @@ where
                 Ok(Some(RuntimeValue::I32(size as i32)))
             }
 
-            GET_CALL_RESULT_FUNC_INDEX => {
+            FunctionIndex::GetCallResultFuncIndex => {
                 // args(0) = pointer to destination in Wasm memory
                 let dest_ptr = Args::parse(args)?;
                 self.set_mem_from_buf(dest_ptr)?;
                 Ok(None)
             }
 
-            GET_UREF_FUNC_INDEX => {
+            FunctionIndex::GetURefFuncIndex => {
                 // args(0) = pointer to uref name in Wasm memory
                 // args(1) = size of uref name
                 // args(2) = pointer to destination in Wasm memory
@@ -583,7 +580,7 @@ where
                 Ok(None)
             }
 
-            HAS_UREF_FUNC_INDEX => {
+            FunctionIndex::HasURefFuncIndex => {
                 // args(0) = pointer to uref name in Wasm memory
                 // args(1) = size of uref name
                 let (name_ptr, name_size) = Args::parse(args)?;
@@ -591,7 +588,7 @@ where
                 Ok(Some(RuntimeValue::I32(result)))
             }
 
-            ADD_UREF_FUNC_INDEX => {
+            FunctionIndex::AddURefFuncIndex => {
                 // args(0) = pointer to uref name in Wasm memory
                 // args(1) = size of uref name
                 // args(2) = pointer to destination in Wasm memory
@@ -600,13 +597,13 @@ where
                 Ok(None)
             }
 
-            GAS_FUNC_INDEX => {
+            FunctionIndex::GasFuncIndex => {
                 let gas: u32 = Args::parse(args)?;
                 self.gas(u64::from(gas))?;
                 Ok(None)
             }
 
-            STORE_FN_INDEX => {
+            FunctionIndex::StoreFnIndex => {
                 // args(0) = pointer to function name in Wasm memory
                 // args(1) = size of the name
                 // args(2) = pointer to additional unforgable names
@@ -627,22 +624,17 @@ where
                 Ok(None)
             }
 
-            PROTOCOL_VERSION_FUNC_INDEX => Ok(Some(self.context.protocol_version().into())),
-
-            REVERT_FUNC_INDEX => {
-                // args(0) = status u32
-                let status = Args::parse(args)?;
-
-                Err(self.revert(status))
+            FunctionIndex::ProtocolVersionFuncIndex => {
+                Ok(Some(self.context.protocol_version().into()))
             }
 
-            SEED_FN_INDEX => {
+            FunctionIndex::SeedFnIndex => {
                 let dest_ptr = Args::parse(args)?;
                 self.write_seed(dest_ptr)?;
                 Ok(None)
             }
 
-            IS_VALID_FN_INDEX => {
+            FunctionIndex::IsValidFnIndex => {
                 // args(0) = pointer to value to validate
                 // args(1) = size of value
                 let (value_ptr, value_size) = Args::parse(args)?;
@@ -654,7 +646,12 @@ where
                 }
             }
 
-            _ => panic!("unknown function index"),
+            FunctionIndex::RevertFuncIndex => {
+                // args(0) = status u32
+                let status = Args::parse(args)?;
+
+                Err(self.revert(status))
+            }
         }
     }
 }
@@ -955,7 +952,7 @@ impl Executor<Module> for WasmiExecutor {
 /// Helper function for creating `known_urefs` associating addresses and corresponding `AccessRights`.
 pub fn key_to_tuple(key: Key) -> Option<([u8; 32], Option<AccessRights>)> {
     match key {
-        Key::URef(raw_addr, maybe_rights) => Some((raw_addr, maybe_rights)),
+        Key::URef(uref) => Some((uref.addr(), uref.access_rights())),
         Key::Account(_) => None,
         Key::Hash(_) => None,
         Key::Local { .. } => None,
@@ -966,7 +963,10 @@ pub fn key_to_tuple(key: Key) -> Option<([u8; 32], Option<AccessRights>)> {
 mod tests {
     use super::Error;
     use common::key::Key;
-    use common::value::account::{AccountActivity, AssociatedKeys, BlockTime, PublicKey, Weight};
+    use common::uref::{AccessRights, URef};
+    use common::value::account::{
+        AccountActivity, AssociatedKeys, BlockTime, PublicKey, PurseId, Weight,
+    };
     use common::value::{Account, Value};
     use engine_state::execution_effect::ExecutionEffect;
     use engine_state::execution_result::ExecutionResult;
@@ -1059,6 +1059,7 @@ mod tests {
                     pub_key,
                     1,
                     BTreeMap::new(),
+                    PurseId::new(URef::new([0u8; 32], AccessRights::READ_ADD_WRITE)),
                     AssociatedKeys::new(PublicKey::new(pub_key), Weight::new(1)),
                     Default::default(),
                     AccountActivity::new(BlockTime(0), BlockTime(0)),