+++ execution-engine/shared/src/init/mod.rs
@@ -1,7 +1,10 @@
 use std::collections::btree_map::BTreeMap;
 
 use common::key::Key;
-use common::value::account::{AccountActivity, AssociatedKeys, BlockTime, PublicKey, Weight};
+use common::uref::{AccessRights, URef};
+use common::value::account::{
+    AccountActivity, AssociatedKeys, BlockTime, PublicKey, PurseId, Weight,
+};
 use common::value::{Account, Value};
 
 pub fn mocked_account(account_addr: [u8; 32]) -> Vec<(Key, Value)> {
@@ -16,6 +19,7 @@ pub fn mocked_account(account_addr: [u8; 32]) -> Vec<(Key, Value)> {
         account_addr,
         0,
         BTreeMap::new(),
+        PurseId::new(URef::new([0u8; 32], AccessRights::READ_ADD_WRITE)),
         associated_keys,
         Default::default(),
         AccountActivity::new(BlockTime(0), BlockTime(100)),