+++ integration-testing/test/test_block_propagation.py
@@ -1,6 +1,7 @@
 import threading
 from test.cl_node.client_parser import parse_show_blocks
 from test.cl_node.docker_node import DockerNode
+from test.cl_node.errors import NonZeroExitCodeError
 from typing import List
 import pytest
 import logging
@@ -13,15 +14,17 @@ from .cl_node.wait import wait_for_blocks_count_at_least, wait_for_peers_count_a
 
 
 class DeployThread(threading.Thread):
-    def __init__(self, node: DockerNode, batches_of_contracts: List[List[str]]) -> None:
+    def __init__(self,
+            node: DockerNode,
+            batches_of_contracts: List[List[str]],
+            max_attempts: int,
+            retry_seconds: int) -> None:
         threading.Thread.__init__(self)
         self.node = node
         self.batches_of_contracts = batches_of_contracts
         self.deployed_blocks_hashes = set()
-
-    def propose(self):
-        propose_output = self.node.client.propose()
-        return extract_block_hash_from_propose_output(propose_output)
+        self.max_attempts = max_attempts
+        self.retry_seconds = retry_seconds
 
     def run(self) -> None:
         for batch in self.batches_of_contracts:
@@ -30,9 +33,10 @@ class DeployThread(threading.Thread):
                                                             payment_contract = contract,
                                                             private_key="validator-0-private.pem",
                                                             public_key="validator-0-public.pem")
-            block_hash = self.propose()
-            self.deployed_blocks_hashes.add(block_hash)
 
+            propose_output = self.node.client.propose_with_retry(self.max_attempts, self.retry_seconds)
+            block_hash = extract_block_hash_from_propose_output(propose_output)
+            self.deployed_blocks_hashes.add(block_hash)
 
 
 @pytest.fixture()
@@ -47,16 +51,17 @@ def nodes(docker_client_fixture):
 
 
 @pytest.mark.parametrize("contract_paths, expected_number_of_blocks", [
-                         ([['test_helloname.wasm'],['test_helloworld.wasm']], 5),
+                         ([['test_helloname.wasm'],['test_helloworld.wasm']], 7),
                          ])
 def test_block_propagation(nodes,
-                           contract_paths: List[List[str]], expected_number_of_blocks):
+                           contract_paths: List[List[str]],
+                           expected_number_of_blocks):
     """
     Feature file: consensus.feature
     Scenario: test_helloworld.wasm deploy and propose by all nodes and stored in all nodes blockstores
     """
 
-    deploy_threads = [DeployThread(node, contract_paths) for node in nodes]
+    deploy_threads = [DeployThread(node, contract_paths, max_attempts = 5, retry_seconds = 3) for node in nodes]
 
     for t in deploy_threads:
         t.start()
@@ -77,11 +82,12 @@ def test_block_propagation(nodes,
 
 
 
-def deploy_and_propose(node, contract):
+def deploy_and_propose(node, contract, nonce=None):
     assert 'Success' in node.client.deploy(session_contract = contract,
                                            payment_contract = contract,
                                            private_key="validator-0-private.pem",
-                                           public_key="validator-0-public.pem")
+                                           public_key="validator-0-public.pem",
+                                           nonce = nonce)
     propose_output = node.client.propose()
     return extract_block_hash_from_propose_output(propose_output)
 
@@ -149,18 +155,18 @@ def test_network_partition_and_rejoin(four_nodes_network):
 
     # Propose separately in each partition. They should not see each others' blocks,
     # so everyone has the genesis plus the 1 block proposed in its partition.
-    deploy_and_propose(partitions[0][0], C[0])
-    deploy_and_propose(partitions[1][0], C[1])
+    # Using the same nonce in both partitions because otherwise one of them will
+    # sit there unable to propose; should use separate accounts really.
+    deploy_and_propose(partitions[0][0], C[0], nonce = 1)
+    deploy_and_propose(partitions[1][0], C[1], nonce = 1)
 
     for node in nodes:
         wait_for_blocks_count_at_least(node, 2, 2, node.timeout * 2)
 
     logging.info("CONNECT PARTITIONS")
-    #four_nodes_network.connect((p[0] for p in partitions))
     for connection in connections_between_partitions:
         four_nodes_network.connect(connection)
 
-    #for node in nodes: wait_for_peers_count_at_least(node, n-1, node.timeout)
     logging.info("PARTITIONS CONNECTED")
 
     # NOTE: Currently `NodeDiscoveryImpl.alivePeersInAscendingDistance` pings the peers
@@ -174,17 +180,17 @@ def test_network_partition_and_rejoin(four_nodes_network):
     # however, nodes in partition[0] will still not see blocks from partition[1]
     # until they also propose a new one on top of the block the created during
     # the network outage.
-    deploy_and_propose(nodes[0], C[2])
+    deploy_and_propose(nodes[0], C[2], nonce = 2)
 
     # NOTE: Currently the node closes the channel only after it has detected a failure,
     # but the syncing will not retry it. A second attempt will create a new channel though.
-    deploy_and_propose(nodes[0], C[2])
+    deploy_and_propose(nodes[0], C[2], nonce = 3)
 
     for node in partitions[0]:
-        logging.info(f"CHECK {node} HAS ALL BLOCKS")
+        logging.info(f"CHECK {node} HAS ALL BLOCKS CREATED IN PARTITION 1")
         wait_for_blocks_count_at_least(node, 4, 4, node.timeout * 2)
 
     for node in partitions[1]:
-        logging.info(f"CHECK {node} HAS ALL NODES")
+        logging.info(f"CHECK {node} HAS ALL BLOCKS CREATED IN PARTITION 1 and 2")
         wait_for_blocks_count_at_least(node, 5, 5, node.timeout * 2)